git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@10561 01821946-851b-0410-b7f4-f1477c70e57a

+++ objects/objectio.cpp
@@ -242,7 +242,7 @@ switch (info.renderType) {
 		rType.lightningInfo.nLength = cf.ReadInt ();
 		rType.lightningInfo.nAmplitude = cf.ReadInt ();
 		rType.lightningInfo.nOffset = cf.ReadInt ();
-		rType.lightningInfo.nWayPoint = (gameData.segs.nLevelVersion <= 22) ? -1 : cf.ReadByte ();
+		rType.lightningInfo.nWayPoint = (gameData.segs.nLevelVersion <= 22) ? -1 : cf.ReadInt ();
 		rType.lightningInfo.nBolts = cf.ReadShort ();
 		rType.lightningInfo.nId = cf.ReadShort ();
 		rType.lightningInfo.nTarget = cf.ReadShort ();
